fitch: implemented the basic functionality

+++ stats/dto/src/main/java/ru/practicum/dto/ViewStatsDto.java
@@ -11,5 +11,5 @@ import lombok.*;
 public class ViewStatsDto {
     private String app;
     private String uri;
-    private Long hits;
+    private Integer hits;
 }
